feat(Synctl::SshConnection): add an ssh input/output object
Add SshConnection which is an improved version of what is returned by
SshProtocol->connect(). The communication model is not synchronous
anymore : each message is tagged with an arbitrary string, both the
client and the server can define hooks to handle messages with a
specific tag. A message can also contain an indication about what tag
to put on the answer so it can be handled.