refactor(Synctl::Ssh): add a multiprotocol Ssh tree
Add a negociation system between ssh client and servers to decide what
version of the ssh protocol to use.

It is unavoidable the ssh protocol will be modified, so at the
beginning of each ssh communication, the client and the server try to
find a common version they can both use. This negociation is made with
a very simple protocol which should not be modified but could be
extended later.

Thus, both client and server should have the code of several ssh
protocol version, so the source tree has been refactored as well as
the test tree.