perf(SshDeposit): add a send-mostly cache to SshDeposit
Add a cache to the SshDeposit allowing to perform send only (with no
need to wait the server answer) operations most of the time in a
normal sending operation.
This reduces transfert time when the RTT is hight, however, since the
communication is made through TCP, even a send only can be blocking if
the server is under hight pressure.