feat(status): add a status command
Add a status command to the main script. The status logic is
implemented in a dedicated object Synctl::Util::Status.
The status command display how files and directory can be treated by
a send or a receive command.
For now it only displays if a file is included or excluded.